endpoint retrieves all trains from the database.
