+++ src/main/java/com/example/trainspottingjava/sighting/Sighting.java
@@ -1,7 +1,7 @@
 package com.example.trainspottingjava.sighting;
 
 import com.example.trainspottingjava.station.Station;
-import com.example.trainspottingjava.train.Train;
+import com.example.trainspottingjava.train.model.Train;
 
 public class Sighting {
     private Station station;
