--- src/main/java/com/example/trainspottingjava/train/TrainController.java
@@ -1,4 +1,0 @@
-package com.example.trainspottingjava.train;
-
-public class TrainController {
-}
