--- src/main/java/com/example/trainspottingjava/train/TrainRepository.java
@@ -1,4 +1,0 @@
-package com.example.trainspottingjava.train;
-
-public class TrainRepository {
-}
