+++ src/main/java/com/example/trainspottingjava/train/controller/TrainController.java
@@ -1,0 +1,28 @@
+package com.example.trainspottingjava.train.controller;
+
+import com.example.trainspottingjava.train.model.Train;
+import com.example.trainspottingjava.train.repository.TrainRepository;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.GetMapping;
+import java.util.ArrayList;
+import java.util.List;
+
+@RestController
+public class TrainController {
+
+    private final TrainRepository trainRepository;
+
+    private List<Train> trains;
+
+    public TrainController(TrainRepository trainRepository){
+        this.trainRepository = trainRepository;
+        this.trains = new ArrayList<>();
+
+    }
+
+    @GetMapping("/Trains")
+    public List<Train> getAllTrains(){
+        trains = trainRepository.findAllByOrderByName().buffer().blockFirst();
+        return trains;
+    }
+}
