+++ src/main/resources/key.json
@@ -1,0 +1,13 @@
+{
+  "type": "service_account",
+  "project_id": "large-fruit-company",
+  "private_key_id": "7bb89dd189caa9314f9f813a64a3f720340c6925",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQDKII86dnCNNgqT\n/Hbsky0fA4EiZU7GpEMNF3l3lbOynk7MsE5Mn0YXnbeyWYUcPP12zDPITqDD2aQJ\nyNkKLLH8L7W0/j8SfjVW9+vf1rnj5sKGadfq3V3CK4P5pMqfYEx/JwvEbajozFxF\nnTXuAQraJsxSQFb1h/7TqYZgMjjoJhTR7/LYf5f6EsnSS/o6lbyLoQwUGoXlieKo\n2IQhzQ2+40W6Kg18DH54mUrQqWECzoOEziqL+r7RlrKxtv12oRwd9lwMus5th9bR\nU56ME/ky/qBhb7+IPIx7ruG39QeIxRxcI1fJQfhLVkZP40V1S1aRLl5alIyydq5O\n+LrFc0FzAgMBAAECggEAFFXhQxb+bN6WgtwAFP4bqIagO9cobUjCbDND57maeip+\nMA6+2eueUCrwe9OQ1k32jDiwLwNmHj4ryEY6O9FSWNZt+BcqPlCPSkAoLX17d0fU\no1RLwlfiInJhVUD0l8NG/vXtM6Z6u20wDVrrPZZCQr9riAzsEsjLZP0lxVDYM9bP\nNNJatZTlkAgVfSrF93na5xrXAgWLO1tpARiycsoxzeMSiJPLey8J+sqwDkkkxBjD\nsYnKUSGagAmfvCKo2A4e83lZE72Pv7ECSK2oTxnISdsUnqd+aT9sssBAtF7SdFBE\nVapiI4x7UyvqoU/G0AQDj6AQg4D/FER6gKwGnUd7nQKBgQDj11JA9eLS0RCiP00n\n1npWYvXXoHctiFbwjtzTcYSHMmBdEDxUqf971qqSTunqCSkhGtKJbyy2hsIiYtpN\nHShO56razI5GjirHDZfhfMjOeKOHySktwAMKTnex3OyZCQWNjELMXoNfdoLbGQ4Y\nwdd6ZAJnarBt4SYJaxqFAVGLvwKBgQDjG6yQZnE9kBEQIgVH/FTstlryafLtykt9\nPBCOt10PmhX8OfD7J+ALFXP/TqoWxi5LIEC05cCh3RBnUSgq6ktu3GafeS9JrsM7\nxEd5ZSxmyV4CTK1DEnxoco+YgbFo4iT8Pv0vu1LNGtHpi6bokbgr8HZUpCkNjJcu\nQEXnSWYHTQKBgB1IZErzD0b3FcM0jY8MMsb9iwk+l3KjrkDWq60FTvotUFNFretX\nuf1JlrydJdUtfjrB/sMpJb8DzNmsiNAWakLR5ZSfa73HsdnJdnbFme1ndeUSxpcZ\nsYpWkW2rrkDw9uS9XSx1DT9stN7B9pzZ9P7G8cQm4ywMHt5hBJSJmdhzAoGAMfIr\nIFK3Qs/FQOQKf0taLfjoxXlRiMQ9l9w20/w1hjHf9emUOtgJKMbeDN7y/zE0iycb\ns+z3UWG5ss8hY1bFoirsktPm93Bs1V8UTPlYK37LovvT+vU80Qeo1uRm7OcXNTZx\nHic3TY5Bbfhdi7ZIZ7lpfLHhqW4q7WwLM5Nr0d0CgYBHdLIxfJ6ZnC5uhzgf0Ki/\nJ5Yhg82AKNnM0gl+rOAylUcjoMv9nbXxvFUV9BZ7WPJ23Bk1wvdjIgAMUgHMsJ0d\nPKfWbstPLLdb8SmkHlnQbD9Cm9zUAMoECb8o3sIHEzWPzHnyuucQwiAzobwmLhZG\nCS+X5YfgGVj+lS1tQAOTBA==\n-----END PRIVATE KEY-----\n",
+  "client_email": "firebase-adminsdk-zy8dr@large-fruit-company.iam.gserviceaccount.com",
+  "client_id": "117197315830388601869",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-zy8dr%40large-fruit-company.iam.gserviceaccount.com",
+  "universe_domain": "googleapis.com"
+}
